+++ rtcclock/rtcclock.h
@@ -1,10 +1,11 @@
 #ifndef RTCCLOCK_H
 #define RTCCLOCK_H
 
+#include<regs/rcc.h>
 #include<regs/rtc.h>
 
 void rtc_init();
 
-uint32_t get_now_micros();
+uint64_t get_now_micros();
 
 #endif
